pyproject.toml
[tool.poetry]
name = "auth-todo"
version = "0.1.0"
description = ""
authors = ["Valerolactone <varanchukv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.34" }
psycopg2-binary = "^2.9.9"
pytest = "^8.2.2"
asyncpg = "^0.29.0"
psycopg = { extras = ["binary"], version = "^3.2.1" }
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
fastapi-mail = "^1.4.1"
pytest-asyncio = "0.24.0"
httpx = "^0.27.2"
alembic = "^1.13.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
black = "^24.4.2"
isort = "^5.13.2"
pylint = "^3.2.4"
flake8 = "^7.1.0"
mypy = "^1.10.1"

[tool.pytest.ini_options]
pythonpath = [
    '.'
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
